requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
torch
huggingface-hub==0.19.4
reportlab==4.0.7
numpy==1.26.2
